References #6732 - predefined list - csv parse fixes

+++ database/migrations/2019_04_25_163655_create_predefined_list_table.php
@@ -14,7 +14,7 @@ class CreatePredefinedListTable extends Migration
     public function up()
     {
         Schema::create('predefined_list', function (Blueprint $table) {
-            $table->integer('eik')->primary();
+            $table->bigInteger('eik')->primary();
             $table->bigInteger('reg_number');
             $table->timestamp('reg_date');
             $table->string('name', 255);
